backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Cors
PyJWT
python-dotenv
